requirements.txt
aws-cdk-lib==2.67.0
constructs>=10.0.0,<11.0.0
PyYAML==6.0
boto3>=1.26.95
cryptography==39.0.2